pyproject.toml
[project]
name = "crewai-veille-simple"
version = "1.0.0"
description = "Système de veille automatisée ultra-simple avec CrewAI - Configuration 100% YAML"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.177.0",
    "crewai-tools>=0.69.0",
    "pyyaml>=6.0",
]

[project.scripts]
veille = "main:main"
